refactor: share same root style
