+++ src/pages/content/index.tsx
@@ -2,7 +2,7 @@ import { addMessageListener, ContentMessage } from "@lib/browser";
 import { parseDocument } from "@lib/readbility";
 import { createRoot } from "react-dom/client";
 
-import "./style.css";
+import "../style.css";
 import Root from "./root";
 
 const addEventListeners = () => {
