--- src/pages/content/style.css
@@ -1,3 +1,0 @@
-@import "tailwindcss" prefix(swc);
-
-@plugin "daisyui";
